requirements.txt
allure-pytest==2.13.5
email_validator==2.2.0
playwright==1.47.0
pydantic==2.9.2
pydantic-settings==2.6.0
pytest==8.3.3
pytest-playwright==0.5.2
pytest-rerunfailures==14.0
pytest-xdist==3.6.1